+++ libs/RSSKit/RSSFeed.m
@@ -2,6 +2,7 @@
  *
  *  GNUstep RSS Kit
  *  Copyright (C) 2006 Guenther Noack
+ *                2010-2012 Free Software Foundation, Inc
  *
  *  This library is free software; you can redistribute it and/or
  *  modify it under the terms of the GNU Lesser General Public
@@ -107,7 +108,7 @@
  */
 -(void) newArticleFound: (id) anArticle
 {
-  int oldArticleIdx;
+  NSUInteger oldArticleIdx;
   NSAssert([articles isKindOfClass: [NSMutableArray class]], @"articles not mutable!");
   
   oldArticleIdx = [articles indexOfObject: anArticle];
